refactor(node): isolate txindex query work and unify accounting
